Move Google Drive code to Android top level
