+++ collect_app/src/main/java/org/odk/collect/android/gdrive/GoogleAccountsManager.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package org.odk.collect.android.utilities.gdrive;
+package org.odk.collect.android.gdrive;
 
 import android.accounts.Account;
 import android.app.Activity;
