+++ collect_app/src/main/java/org/odk/collect/android/preferences/ServerPreferencesFragment.java
@@ -51,7 +51,7 @@ import org.odk.collect.android.utilities.PlayServicesChecker;
 import org.odk.collect.android.utilities.SoftKeyboardUtils;
 import org.odk.collect.android.utilities.ToastUtils;
 import org.odk.collect.android.utilities.Validator;
-import org.odk.collect.android.utilities.gdrive.GoogleAccountsManager;
+import org.odk.collect.android.gdrive.GoogleAccountsManager;
 
 import java.io.ByteArrayInputStream;
 import java.util.Locale;
