+++ collect_app/src/test/java/org/odk/collect/android/gdrive/GoogleAccountsManagerTest.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.utilities.gdrive;
+package org.odk.collect.android.gdrive;
 
 import android.accounts.Account;
 import android.content.Intent;
